+++ tests/testthat/test-append-debug.R
@@ -24,4 +24,15 @@ with_test_authentication({
         expect_identical(as.vector(out$C),
             factor(c(rep(NA, 1000), rep(c("C", "D"), 500))))
     })
+
+    test_that("Can re-alias array variables to make them line up (and old refs don't reappear)", {
+        part1 <- mrdf.setup(newDataset(mrdf), name="CA1")
+        part2 <- mrdf.setup(newDataset(mrdf), name="CA2")
+        expect_identical(aliases(subvariables(part1$CA1)),
+            aliases(subvariables(part2$CA2)))
+        alias(part2$CA2) <- "CA"
+        alias(part1$CA1) <- "CA"
+        out <- appendDataset(part1, part2)
+        expect_equal(dim(out), c(2*nrow(part2), ncol(part2)))
+    })
 })